Added Simple yet Buggy Camera

+++ FabledRealms/src/Engine/Time.h
@@ -3,5 +3,5 @@
 struct Time
 {
 	double currentTime = 0.0;
-	double deltaTime = 0.0;
+	float deltaTime = 0.0;
 };
